pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "factsumm"
version = "1.2.2.post1"
description = "Towards GPT-4 like large language and visual assistant."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "torch==2.1.2", "torchvision==0.16.2",
    "transformers==4.37.2", "tokenizers==0.15.1", "sentencepiece==0.1.99", 
    "absl-py==2.1.0", "accelerate==0.27.2", "aiohttp==3.9.3", "aiosignal==1.3.1", "annotated-types==0.6.0", "anyio==4.2.0", "appdirs==1.4.4", "argon2-cffi==23.1.0", "argon2-cffi-bindings==21.2.0", "arrow==1.3.0", "asttokens==2.4.1", "async-lru==2.0.4", "async-timeout==4.0.3", "attrs==23.2.0", "Babel==2.14.0", "beautifulsoup4==4.12.3", "bitsandbytes==0.42.0", "bleach==6.1.0", "blis==0.7.11", "cachetools==5.3.2", "catalogue==2.0.10", "certifi==2024.2.2", "cffi==1.16.0", "charset-normalizer==3.3.2", "click==8.1.7", "cloudpathlib==0.16.0", "comm==0.2.1", "confection==0.1.4", "contourpy==1.2.0", "cycler==0.12.1", "cymem==2.0.8", "dataclasses-json==0.6.4", "datasets==2.17.1", "debugpy==1.8.1", "decorator==5.1.1", "deepspeed==0.13.2", "defusedxml==0.7.1", "dill==0.3.8", "distro==1.9.0", "docker-pycreds==0.4.0", "en-core-web-sm==3.7.1", "et-xmlfile==1.1.0", "evaluate==0.4.1", "Events==0.5", "exceptiongroup==1.2.0", "executing==2.0.1", "factsumm==1.2.2.post1", "fastjsonschema==2.19.1", "filelock==3.13.1", "fonttools==4.49.0", "fqdn==1.5.1", "frozenlist==1.4.1", "fsspec==2023.10.0", "gevent==23.9.1", "gevent-websocket==0.10.1", "gitdb==4.0.11", "GitPython==3.1.42", "google-auth==2.27.0", "google-auth-oauthlib==1.2.0", "greenlet==3.0.3", "grpcio==1.60.1", "h11==0.14.0", "haven-ai==0.7.3", "hjson==3.1.0", "httpcore==1.0.2", "httpx==0.26.0", "huggingface-hub==0.20.3", "idna==3.6", "ipykernel==6.29.2", "ipython==8.21.0", "isoduration==20.11.0", "jedi==0.19.1", "Jinja2==3.1.3", "joblib==1.3.2", "json5==0.9.14", "jsonpatch==1.33", "jsonpointer==2.4", "jsonschema==4.21.1", "jsonschema-specifications==2023.12.1", "jupyter_client==8.6.0", "jupyter_core==5.7.1", "jupyter-events==0.9.0", "jupyter-lsp==2.2.2", "jupyter_server==2.12.5", "jupyter_server_terminals==0.5.2", "jupyterlab==4.0.12", "jupyterlab_pygments==0.3.0", "jupyterlab_server==2.25.2", "kiwisolver==1.4.5", "langchain==0.1.7", "langchain-community==0.0.20", "langchain-core==0.1.23", "langchain-openai==0.0.6", "langchainhub==0.1.14", "langcodes==3.3.0",
    "langsmith==0.0.87", "lightning-utilities==0.10.1", "Markdown==3.5.2", "MarkupSafe==2.1.5", "marshmallow==3.20.2", "matplotlib==3.8.3", "matplotlib-inline==0.1.6", "mistune==3.0.2", "mpmath==1.3.0", "multidict==6.0.5", "multiprocess==0.70.16", "murmurhash==1.0.10", "mypy-extensions==1.0.0", "nbclient==0.9.0", "nbconvert==7.16.0", "nbformat==5.9.2", "nest-asyncio==1.6.0", "netkit==3.1.12", "networkx==3.2.1", "ninja==1.11.1.1", "nltk==3.8.1", "notebook==7.0.8", "notebook_shim==0.2.3", "numpy==1.26.4", "nvidia-cublas-cu12==12.1.3.1", "nvidia-cuda-cupti-cu12==12.1.105", "nvidia-cuda-nvrtc-cu12==12.1.105", "nvidia-cuda-runtime-cu12==12.1.105", "nvidia-cudnn-cu12==8.9.2.26", "nvidia-cufft-cu12==11.0.2.54", "nvidia-curand-cu12==10.3.2.106", "nvidia-cusolver-cu12==11.4.5.107", "nvidia-cusparse-cu12==12.1.0.106", "nvidia-nccl-cu12==2.18.1", "nvidia-nvjitlink-cu12==12.3.101", "nvidia-nvtx-cu12==12.1.105", "oauthlib==3.2.2", "openai==1.12.0", "openpyxl==3.1.2", "overrides==7.7.0", "packaging==23.2", "pandas==2.2.0", "pandocfilters==1.5.1", "parso==0.8.3", "peft==0.8.2", "pexpect==4.9.0", "pillow==10.2.0", "pip==24.0", "platformdirs==4.2.0", "preshed==3.0.9", "prometheus-client==0.19.0", "prompt-toolkit==3.0.43", "protobuf==3.20.1", "psutil==5.9.8", "ptyprocess==0.7.0", "pure-eval==0.2.2", "py-cpuinfo==9.0.0", "pyarrow==15.0.0", "pyarrow-hotfix==0.6", "pyasn1==0.5.1", "pyasn1-modules==0.3.0", "pycparser==2.21", "pydantic==2.6.1", "pydantic_core==2.16.2", "pyDeprecate==0.3.2", "Pygments==2.17.2", "pynvml==11.5.0", "pyparsing==3.1.1", "python-dateutil==2.8.2", "python-json-logger==2.0.7", "pytorch-lightning==1.6.5", "pytorch-ranger==0.1.1", "pytz==2024.1", "PyYAML==6.0.1", "pyzmq==25.1.2", "referencing==0.33.0", "regex==2023.12.25", "requests==2.31.0", "requests-oauthlib==1.3.1", "responses==0.18.0", "rfc3339-validator==0.1.4", "rfc3986-validator==0.1.1", "rouge-score==0.1.2", "rpds-py==0.17.1", "rsa==4.9", "safetensors==0.4.2", "scikit-learn==1.4.0",
    "scipy==1.12.0", "Send2Trash==1.8.2", "sentence-transformers==2.3.1", "sentencepiece==0.1.99", "sentry-sdk==1.40.6", "setproctitle==1.3.3", "setuptools==68.2.2", "six==1.16.0", "smart-open==6.4.0", "smmap==5.0.1", "sniffio==1.3.0", "soupsieve==2.5", "spacy==3.7.3", "spacy-legacy==3.0.12", "spacy-loggers==1.0.5", "SQLAlchemy==2.0.27", "srsly==2.4.8", "stack-data==0.6.3", "sympy==1.12", "tavily-python==0.3.1", "tenacity==8.2.3", "tensorboard==2.15.1", "tensorboard-data-server==0.7.2", "terminado==0.18.0", "thinc==8.2.3", "threadpoolctl==3.2.0", "tiktoken==0.5.2", "tinycss2==1.2.1", "tokenizers==0.15.1", "tomli==2.0.1", "torch==2.1.2", "torch-optimizer==0.3.0", "torchmetrics==1.3.0.post0", "torchvision==0.16.2", "tornado==6.4", "tqdm==4.66.1", "traitlets==5.14.1", "transformers==4.37.2", "triton==2.1.0", "typer==0.9.0", "types-python-dateutil==2.8.19.20240106", "types-requests==2.31.0.20240218", "typing_extensions==4.9.0", "typing-inspect==0.9.0", "tzdata==2024.1", "uri-template==1.3.0", "urllib3==2.2.0", "wandb==0.16.3", "wasabi==1.1.2", "wcwidth==0.2.13", "weasel==0.3.4", "webcolors==1.13", "webencodings==0.5.1", "websocket-client==1.7.0", "Werkzeug==3.0.1", "wheel==0.41.2", "xxhash==3.4.1", "yarl==1.9.4", "zope.event==5.0", "zope.interface==6.1"
]

[project.optional-dependencies]
train = ["wandb"]
build = ["build", "twine"]

[project.urls]
"Homepage" =  "http://www.google.com"
"Bug Tracker" =  "http://www.google.com"

[tool.setuptools.packages.find]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]

[tool.wheel]
exclude = ["assets*", "benchmark*", "docs", "dist*", "playground*", "scripts*", "tests*"]